# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HOFA Backend API",
        "description": "API documentation for HOFA Backend application",
        "contact": {
            "name": "Softskills Support",
            "email": "direction@softskills.ci"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/dons/{enfiler}": {
            "get": {
                "tags": [
                    "Dons"
                ],
                "summary": "Récupérer un don spécifique",
                "description": "Récupérer les détails d'un don par son ID",
                "operationId": "showDon",
                "parameters": [
                    {
                        "name": "enfiler",
                        "in": "path",
                        "description": "ID du don",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01jywbsemp4vdwx02h17z5mgah"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Don récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Enfiler retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "01jywbsemp4vdwx02h17z5mgah"
                                                        },
                                                        "slug": {
                                                            "type": "string",
                                                            "example": "ENF-12345678-1234-1234-1234-123456789abc"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Jean Dupont"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "jean@example.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "+33123456789",
                                                            "nullable": true
                                                        },
                                                        "type_enfiler": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "string",
                                                                    "example": "01jywbsemp4vdwx02h17z5mgah"
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Don financier"
                                                                },
                                                                "description": {
                                                                    "type": "string",
                                                                    "example": "Don en argent"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "is_active": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-06-28T22:35:27.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-06-28T22:35:27.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Don non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Don not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dons"
                ],
                "summary": "Désactiver un don",
                "description": "Supprimer un don en définissant is_active à false",
                "operationId": "desactivateDon",
                "parameters": [
                    {
                        "name": "enfiler",
                        "in": "path",
                        "description": "ID du don",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01jywbsemp4vdwx02h17z5mgah"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Don désactivé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Enfiler deleted successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Don non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Don not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Utilisez un jeton JWT pour l'authentification.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}